+++ src/pg/index.js
@@ -1,9 +1,12 @@
 const utils = require('../utils');
 
 async function registerType(client) {
-  const result = await client.query('SELECT typname, oid FROM pg_type WHERE typname IN ($1, $2)', ['vector', 'halfvec']);
-  const vector = result.rows.find((v) => v.typname == 'vector');
-  const halfvec = result.rows.find((v) => v.typname == 'halfvec');
+  const result = await client.query('SELECT typname, oid FROM pg_type WHERE typname IN ($1, $2, $3)', ['vector', 'halfvec', 'sparsevec']);
+  const rows = result.rows;
+
+  const vector = rows.find((v) => v.typname == 'vector');
+  const halfvec = rows.find((v) => v.typname == 'halfvec');
+  const sparsevec = rows.find((v) => v.typname == 'sparsevec');
 
   if (!vector) {
     throw new Error('vector type not found in the database');
@@ -16,6 +19,12 @@ async function registerType(client) {
   if (halfvec) {
     client.setTypeParser(halfvec.oid, 'text', function(value) {
       return utils.fromSql(value);
+    });
+  }
+
+  if (sparsevec) {
+    client.setTypeParser(sparsevec.oid, 'text', function(value) {
+      return utils.sparsevecFromSql(value);
     });
   }
 }