+++ tests/pg/index.test.mjs
@@ -1,5 +1,6 @@
 import pg from 'pg';
 import pgvector from 'pgvector/pg';
+import { SparseVector } from 'pgvector';
 
 test('example', async () => {
   const client = new pg.Client({database: 'pgvector_node_test'});
@@ -9,15 +10,15 @@ test('example', async () => {
   await pgvector.registerType(client);
 
   await client.query('DROP TABLE IF EXISTS pg_items');
-  await client.query('CREATE TABLE pg_items (id serial PRIMARY KEY, embedding vector(3), half_embedding halfvec(3), binary_embedding bit(3))');
+  await client.query('CREATE TABLE pg_items (id serial PRIMARY KEY, embedding vector(3), half_embedding halfvec(3), binary_embedding bit(3), sparse_embedding sparsevec(3))');
 
   const params = [
-    pgvector.toSql([1, 1, 1]), pgvector.toSql([1, 1, 1]), '000',
-    pgvector.toSql([2, 2, 2]), pgvector.toSql([2, 2, 2]), '101',
-    pgvector.toSql([1, 1, 2]), pgvector.toSql([1, 1, 2]), '111',
-    null, null, null
+    pgvector.toSql([1, 1, 1]), pgvector.toSql([1, 1, 1]), '000', SparseVector.fromDense([1, 1, 1]).toSql(),
+    pgvector.toSql([2, 2, 2]), pgvector.toSql([2, 2, 2]), '101', SparseVector.fromDense([2, 2, 2]).toSql(),
+    pgvector.toSql([1, 1, 2]), pgvector.toSql([1, 1, 2]), '111', SparseVector.fromDense([1, 1, 2]).toSql(),
+    null, null, null, null
   ];
-  await client.query('INSERT INTO pg_items (embedding, half_embedding, binary_embedding) VALUES ($1, $2, $3), ($4, $5, $6), ($7, $8, $9), ($10, $11, $12)', params);
+  await client.query('INSERT INTO pg_items (embedding, half_embedding, binary_embedding, sparse_embedding) VALUES ($1, $2, $3, $4), ($5, $6, $7, $8), ($9, $10, $11, $12), ($13, $14, $15, $16)', params);
 
   const { rows } = await client.query('SELECT * FROM pg_items ORDER BY embedding <-> $1 LIMIT 5', [pgvector.toSql([1, 1, 1])]);
   expect(rows.map(v => v.id)).toStrictEqual([1, 3, 2, 4]);
@@ -30,6 +31,9 @@ test('example', async () => {
   expect(rows[0].binary_embedding).toStrictEqual('000');
   expect(rows[1].binary_embedding).toStrictEqual('111');
   expect(rows[2].binary_embedding).toStrictEqual('101');
+  expect(rows[0].sparse_embedding.toArray()).toStrictEqual([1, 1, 1]);
+  expect(rows[1].sparse_embedding.toArray()).toStrictEqual([1, 1, 2]);
+  expect(rows[2].sparse_embedding.toArray()).toStrictEqual([2, 2, 2]);
 
   await client.query('CREATE INDEX ON pg_items USING hnsw (embedding vector_l2_ops)');